Extract analytics helpers into dedicated module
